requirements.txt
langchain
langchain-openai
langchain_community
python-dotenv
streamlit
langchain_core